WIP: split image proxy refresh fixes for LOC (checkpoint)
