+++ Services/ProxiedImageCacheService.cs
@@ -11,7 +11,7 @@ namespace Wayfarer.Services;
 /// Read operations are lock-free for concurrent performance; writes and eviction
 /// are serialized via static SemaphoreSlim (same pattern as <see cref="TileCacheService"/>).
 /// </summary>
-public class ProxiedImageCacheService : IProxiedImageCacheService
+public partial class ProxiedImageCacheService : IProxiedImageCacheService
 {
     private readonly ILogger<ProxiedImageCacheService> _logger;
     private readonly ApplicationDbContext _dbContext;
@@ -199,144 +199,6 @@ public class ProxiedImageCacheService : IProxiedImageCacheService
     }
 
     /// <inheritdoc />
-    public async Task<ProxiedImageCacheStoreResult> SetAsync(string cacheKey, byte[] bytes, string contentType)
-    {
-        var settings = _settingsService.GetSettings();
-
-        // Caching disabled
-        if (settings.MaxCacheImageSizeInMB < 0)
-            return ProxiedImageCacheStoreResult.Failure;
-
-        var filePath = Path.Combine(_cacheDirectory, $"{cacheKey}.dat");
-        var tempFilePath = CreateTempImagePath(filePath);
-        try
-        {
-            Directory.CreateDirectory(_cacheDirectory);
-            await File.WriteAllBytesAsync(tempFilePath, bytes);
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Error writing proxy image file for key {CacheKey}.", cacheKey);
-            return ProxiedImageCacheStoreResult.Failure;
-        }
-
-        // Lock only for DB operations and size counter mutations
-        await _cacheLock.WaitAsync();
-        try
-        {
-            // Check if entry already exists (race: another request cached it while we were downloading)
-            var existing = await _dbContext.ImageCacheMetadata
-                .FirstOrDefaultAsync(m => m.CacheKey == cacheKey);
-
-            if (existing != null)
-            {
-                return await ReplaceExistingEntryAsync(existing, tempFilePath, bytes, contentType);
-            }
-
-            // Evict in a loop until enough space is available or no more entries remain
-            var maxSizeBytes = settings.MaxCacheImageSizeInMB * 1024L * 1024L;
-            while (Interlocked.Read(ref _currentCacheSize) + bytes.Length > maxSizeBytes)
-            {
-                var evictedCount = await EvictLruEntriesAsync();
-                if (evictedCount == 0) break;
-            }
-
-            // New entries move bytes first, then persist metadata. If metadata fails, the
-            // unreferenced file is deleted and callers see a failed store.
-            var metadata = new ImageCacheMetadata
-            {
-                CacheKey = cacheKey,
-                ContentType = contentType,
-                FilePath = filePath,
-                Size = bytes.Length,
-                CreatedAt = DateTime.UtcNow,
-                LastAccessed = DateTime.UtcNow
-            };
-
-            _dbContext.ImageCacheMetadata.Add(metadata);
-            try
-            {
-                ReplaceImageFileAtomically(tempFilePath, filePath);
-                await SaveMetadataChangesAsync();
-                Interlocked.Add(ref _currentCacheSize, bytes.Length);
-            }
-            catch
-            {
-                _dbContext.ImageCacheMetadata.Remove(metadata);
-                try { File.Delete(filePath); } catch { /* best-effort cleanup */ }
-                try { File.Delete(tempFilePath); } catch { /* best-effort cleanup */ }
-                return ProxiedImageCacheStoreResult.Failure;
-            }
-
-            _logger.LogInformation("Cached proxy image: key={CacheKey}, size={Size} bytes.", cacheKey, bytes.Length);
-            return ProxiedImageCacheStoreResult.Success;
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Error caching proxy image for key {CacheKey}.", cacheKey);
-            return ProxiedImageCacheStoreResult.Failure;
-        }
-        finally
-        {
-            TryDeleteTempImage(tempFilePath);
-            _cacheLock.Release();
-        }
-    }
-
-    /// <summary>
-    /// Stores replacement bytes without overwriting the old usable file before metadata succeeds.
-    /// </summary>
-    private async Task<ProxiedImageCacheStoreResult> ReplaceExistingEntryAsync(
-        ImageCacheMetadata existing,
-        string tempFilePath,
-        byte[] bytes,
-        string contentType)
-    {
-        var oldFilePath = existing.FilePath;
-        var oldContentType = existing.ContentType;
-        var oldSize = existing.Size;
-        var oldCreatedAt = existing.CreatedAt;
-        var oldLastAccessed = existing.LastAccessed;
-        var newFilePath = CreateReplacementImagePath(oldFilePath);
-
-        try
-        {
-            // The metadata row is the commit point. New bytes live in an unreferenced
-            // sibling file until the row points at them, so failed metadata leaves the
-            // old file and metadata usable. On post-save cleanup failure, serving still
-            // uses the new metadata and file; only the old file may linger.
-            ReplaceImageFileAtomically(tempFilePath, newFilePath);
-            var now = DateTime.UtcNow;
-            existing.FilePath = newFilePath;
-            existing.ContentType = contentType;
-            existing.Size = bytes.Length;
-            existing.CreatedAt = now;
-            existing.LastAccessed = now;
-
-            var saved = await SaveWithConcurrencyRetryAsync(existing);
-            if (!saved)
-            {
-                RestoreMetadataValues(existing, oldFilePath, oldContentType, oldSize, oldCreatedAt, oldLastAccessed);
-                TryDeleteTempImage(newFilePath);
-                return ProxiedImageCacheStoreResult.Failure;
-            }
-
-            Interlocked.Add(ref _currentCacheSize, bytes.Length - oldSize);
-            TryDeleteTempImage(oldFilePath);
-            _logger.LogInformation("Refreshed proxy image: key={CacheKey}, size={Size} bytes.",
-                existing.CacheKey, bytes.Length);
-            return ProxiedImageCacheStoreResult.Success;
-        }
-        catch (Exception ex)
-        {
-            RestoreMetadataValues(existing, oldFilePath, oldContentType, oldSize, oldCreatedAt, oldLastAccessed);
-            TryDeleteTempImage(newFilePath);
-            _logger.LogError(ex, "Error replacing proxy image file for key {CacheKey}.", existing.CacheKey);
-            return ProxiedImageCacheStoreResult.Failure;
-        }
-    }
-
-    /// <inheritdoc />
     public async Task<double> GetCacheSizeInMbAsync()
     {
         var totalSize = await _dbContext.ImageCacheMetadata.SumAsync(m => (long)m.Size);
@@ -415,141 +277,4 @@ public class ProxiedImageCacheService : IProxiedImageCacheService
         return entriesToEvict.Count;
     }
 
-    /// <summary>
-    /// Saves metadata changes with retry on concurrency conflicts.
-    /// Uses the same retry pattern as <see cref="TileCacheService"/>.
-    /// </summary>
-    private async Task<bool> SaveWithConcurrencyRetryAsync(ImageCacheMetadata metadata)
-    {
-        var attempts = 0;
-        var updated = false;
-
-        while (!updated && attempts < 3)
-        {
-            attempts++;
-            try
-            {
-                _dbContext.ImageCacheMetadata.Update(metadata);
-                await SaveMetadataChangesAsync();
-                updated = true;
-            }
-            catch (DbUpdateConcurrencyException ex)
-            {
-                var entry = ex.Entries.Single();
-                var databaseValues = await entry.GetDatabaseValuesAsync();
-
-                if (databaseValues == null)
-                {
-                    _logger.LogWarning("Image cache metadata was deleted by another process for key {CacheKey}.",
-                        metadata.CacheKey);
-                    return false;
-                }
-
-                // Reload database values and reapply our LastAccessed update
-                var dbMetadata = (ImageCacheMetadata)databaseValues.ToObject();
-                metadata.LastAccessed = DateTime.UtcNow;
-                entry.OriginalValues.SetValues(databaseValues);
-            }
-        }
-
-        return updated;
-    }
-
-    /// <summary>
-    /// Creates a same-directory temporary path for atomic image replacement.
-    /// </summary>
-    private static string CreateTempImagePath(string filePath)
-    {
-        var directory = Path.GetDirectoryName(filePath) ?? ".";
-        var fileName = Path.GetFileName(filePath);
-        return Path.Combine(directory, $"{fileName}.{Guid.NewGuid():N}.tmp");
-    }
-
-    /// <summary>
-    /// Creates a same-directory replacement path that is not referenced until metadata commits.
-    /// </summary>
-    private static string CreateReplacementImagePath(string filePath)
-    {
-        var directory = Path.GetDirectoryName(filePath) ?? ".";
-        var extension = Path.GetExtension(filePath);
-        var baseName = Path.GetFileNameWithoutExtension(filePath);
-        return Path.Combine(directory, $"{baseName}.{Guid.NewGuid():N}{extension}");
-    }
-
-    /// <summary>
-    /// Restores tracked metadata values after an uncommitted replacement failure.
-    /// </summary>
-    private static void RestoreMetadataValues(
-        ImageCacheMetadata metadata,
-        string filePath,
-        string contentType,
-        int size,
-        DateTime createdAt,
-        DateTime lastAccessed)
-    {
-        metadata.FilePath = filePath;
-        metadata.ContentType = contentType;
-        metadata.Size = size;
-        metadata.CreatedAt = createdAt;
-        metadata.LastAccessed = lastAccessed;
-    }
-
-    /// <summary>
-    /// Replaces the final image using the active production or test hook.
-    /// </summary>
-    private static void ReplaceImageFileAtomically(string tempFilePath, string filePath) =>
-        _replaceImageFile(tempFilePath, filePath);
-
-    /// <summary>
-    /// Replaces the final image with a same-directory temp file so readers never see partial bytes.
-    /// </summary>
-    private static void ReplaceImageFileAtomicallyCore(string tempFilePath, string filePath)
-    {
-        if (File.Exists(filePath))
-        {
-            File.Replace(tempFilePath, filePath, null);
-            return;
-        }
-
-        File.Move(tempFilePath, filePath);
-    }
-
-    /// <summary>
-    /// Saves metadata changes using the active production or test hook.
-    /// </summary>
-    private Task<int> SaveMetadataChangesAsync() => _saveMetadataChanges(_dbContext);
-
-    /// <summary>
-    /// Overrides image replacement for deterministic tests.
-    /// </summary>
-    internal static void SetImageFileReplacerForTesting(Action<string, string>? replacer)
-    {
-        _replaceImageFile = replacer ?? ReplaceImageFileAtomicallyCore;
-    }
-
-    /// <summary>
-    /// Overrides metadata persistence for deterministic tests.
-    /// </summary>
-    internal static void SetMetadataSaverForTesting(Func<ApplicationDbContext, Task<int>>? saver)
-    {
-        _saveMetadataChanges = saver ?? (dbContext => dbContext.SaveChangesAsync());
-    }
-
-    /// <summary>
-    /// Deletes an unused temp file without masking the original write or replacement error.
-    /// </summary>
-    private static void TryDeleteTempImage(string tempFilePath)
-    {
-        try
-        {
-            if (File.Exists(tempFilePath))
-            {
-                File.Delete(tempFilePath);
-            }
-        }
-        catch
-        {
-            // Best-effort cleanup only.
-        }
-    }
 }
